♻️ refactor(http): extract customFetch to adapter
- Move fetch adapter to a dedicated file
- Decouple SDK fetch logic from main entry point
